moving risk_data function to a new file and moving Escribir_Datos_Osiris to a new file to avoid circular import between risk_data file  and subida.py
